% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_med.R
\name{flex_med}
\alias{flex_med}
\title{Mediation Analysis with Centred Two-Piece Student \eqn{t} (CTPT) Errors}
\usage{
flex_med(
  X,
  M,
  Y,
  model = "full",
  q00 = 1/3,
  q01 = 1/3,
  alpha_gamma = 2,
  beta_gamma = 2,
  gamma_lower = 0.05,
  gamma_upper = 20,
  d_nu = 0.01,
  nu_lower = 1.01,
  iter = 30000,
  warmup = floor(iter/2),
  chains = 1,
  seed = 123,
  verbose = FALSE,
  refresh = 2500,
  repetitions = 1,
  method = "normal",
  bridgesampling_cores = 1,
  use_neff = TRUE,
  maxiter = 1000,
  silent = FALSE,
  ...
)
}
\arguments{
\item{X}{A numeric vector of length \eqn{n} (the independent variable).}

\item{M}{A numeric vector of length \eqn{n} (the mediator).}

\item{Y}{A numeric vector of length \eqn{n} (the response).}

\item{model}{A string: which model to use ("full", "gamma-only", "nu-only", "normal")}

\item{q00}{A non-negative numeric value giving the conditional probability
\eqn{p(\alpha=0,\beta=0 \mid \alpha\beta=0)}. Defaults to \code{1/3}.}

\item{q01}{A non-negative numeric value giving the conditional probability
\eqn{p(\alpha=0,\beta\neq 0 \mid \alpha\beta=0)}. Defaults to \code{1/3}.

Here, \eqn{q00 + q01 + q10 = 1}, where \eqn{q10 = p(\alpha\neq 0,\beta=0 \mid \alpha\beta=0)}
is computed internally as \code{1 - q00 - q01}. These three probabilities correspond
to the disjoint ways in which the mediation effect can be zero (i.e.,
\eqn{\alpha\beta = 0}). They must be non-negative and sum to 1.}

\item{alpha_gamma, beta_gamma}{Shape and rate for the Gamma prior on skewness parameter \eqn{\gamma}.}

\item{gamma_lower, gamma_upper}{Truncation bounds for \eqn{\gamma} (\eqn{0 < \gamma_{\text{lower}} \leq \gamma \leq \gamma_{\text{upper}}}).}

\item{d_nu}{Rate for the exponential prior on degrees-of-freedom \eqn{\nu}.}

\item{nu_lower}{Lower truncation bound for \eqn{\nu} (must satisfy \eqn{\nu > 1}).}

\item{iter, warmup, chains, seed, verbose, refresh, ...}{Additional arguments passed to \code{rstan::stan()}.}

\item{repetitions, method, bridgesampling_cores, use_neff, maxiter, silent}{Additional arguments passed to \code{bridgesampling::bridge_sampler()}.}
}
\value{
A list with the following elements:
\describe{
\item{\code{X2M_with_alpha}}{A \code{stanfit} object for the mediator model with \eqn{\alpha \neq 0}.}
\item{\code{XM2Y_with_beta}}{A \code{stanfit} object for the outcome model with \eqn{\beta \neq 0}.}
\item{\code{X2M_without_alpha}}{A \code{stanfit} object for the mediator model with \eqn{\alpha = 0}.}
\item{\code{XM2Y_without_beta}}{A \code{stanfit} object for the outcome model with \eqn{\beta = 0}.}
\item{\code{samples_med}}{A numeric vector of posterior draws for \eqn{\alpha \beta}.}
\item{\code{bf_med}}{A numeric value giving the Bayes factor for the mediation effect.}
}
}
\description{
Mediation Analysis with Centred Two-Piece Student \eqn{t} (CTPT) Errors
}
