% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flex_mr.R
\name{flex_mr}
\alias{flex_mr}
\title{Fit a Linear Regression with Centred Two-Piece Student \eqn{t} (CTPT) Errors}
\usage{
flex_mr(
  X,
  Y,
  model = "full",
  alpha_gamma = 2,
  beta_gamma = 2,
  gamma_lower = 0.05,
  gamma_upper = 20,
  d_nu = 0.01,
  nu_lower = 1.01,
  iter = 30000,
  warmup = floor(iter/2),
  chains = 1,
  seed = 123,
  verbose = TRUE,
  refresh = 2500,
  ...
)
}
\arguments{
\item{X}{A numeric matrix of predictors (\eqn{n \times k}).}

\item{Y}{A numeric vector of length \eqn{n} (the response).}

\item{model}{A string: which model to use ("full", "gamma-only", "nu-only", "normal")}

\item{alpha_gamma, beta_gamma}{Shape and rate for the Gamma prior on skewness parameter \eqn{\gamma}.}

\item{gamma_lower, gamma_upper}{Truncation bounds for \eqn{\gamma} (\eqn{0 < \gamma_{\text{lower}} \leq \gamma \leq \gamma_{\text{upper}}}).}

\item{d_nu}{Rate for the exponential prior on degrees-of-freedom \eqn{\nu}.}

\item{nu_lower}{Lower truncation bound for \eqn{\nu} (must satisfy \eqn{\nu > 1}).}

\item{iter, warmup, chains, seed, verbose, refresh, ...}{Additional arguments passed to \code{rstan::stan()}.}
}
\value{
An object of class \code{stanfit}.
}
\description{
Fit a Linear Regression with Centred Two-Piece Student \eqn{t} (CTPT) Errors
}
